fix dispatch terminal finalization edge cases
